ui(#3): 이번주 근무 내역과 근무 시간 막대 컴포넌트 분리
